Moved back testing algorithms to services directory. Added registration to pick algorithm
