+++ services/back_tester_interface.py
@@ -1,0 +1,6 @@
+
+from typing import Dict
+
+class BackTesterInterface:
+    def backtest(self, stocks: Dict[str, float]) -> None:
+        pass
